hide proxydb (venndb) generated code behind wrapper type
